requirements.txt
absl-py==1.4.0
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
comm==0.1.2
contourpy==1.0.6
cryptography==41.0.1
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==3.2.19
django-allauth==0.51.0
django-cors-headers==4.0.0
django-debug-toolbar==4.1.0
django-extensions==3.2.3
django-filter==23.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
entrypoints==0.4
executing==1.2.0
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.38.0
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
ipykernel==6.19.4
ipython==8.8.0
jax==0.4.12
jedi==0.18.2
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==5.1.2
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lmdb==1.4.1
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.6.2
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mysqlclient==2.1.1
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
protobuf==4.23.2
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.14.0
PyJWT==1.7.1
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7
pywin32==305
pyzmq==24.0.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
statsmodels==0.13.5
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.8.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.14
wcwidth==0.2.5
Werkzeug==2.3.6
wrapt==1.14.1
